+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.15.0
+  * Change non-transformer message rewriting to preserve the original message as
+    much as possible. Adds --useStringSubstitution command-line arg.
+
 ## 0.14.0
   * Split message extraction and code generation out into a separate
     package. Versioned to match the corresponding Intl version.